refactor(Provider): simplify API providers querying multiple regions
Provide abstract base classes MetadataApiProvider and ReleaseApiLookup
that hold functionality common to typical API based providers. Move
queryAllRegions helper there and base Tidal, iTunes and Deezer on this.

Adapt iTunes to use queryAllRegions.
